backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pymilvus==2.3.0
python-jose==3.3.0
python-multipart==0.0.6
pydantic==1.10.8
passlib==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1 